APP-2887 Renamed config exception package
