--- symphony-bdk-core/src/main/java/com/symphony/bdk/core/config/BdkConfigException.java
@@ -1,16 +1,0 @@
-package com.symphony.bdk.core.config;
-
-/**
- *
- */
-public class BdkConfigException extends RuntimeException {
-
-  /**
-   *
-   * @param message
-   * @param source
-   */
-  public BdkConfigException(String message, Exception source) {
-    super(message, source);
-  }
-}
