+++ symphony-bdk-core/src/main/java/com/symphony/bdk/core/config/BdkConfigLoader.java
@@ -1,13 +1,14 @@
 package com.symphony.bdk.core.config;
 
+import com.symphony.bdk.core.config.exception.BdkConfigException;
+import com.symphony.bdk.core.config.legacy.LegacyConfigMapper;
+import com.symphony.bdk.core.config.legacy.model.LegacySymConfig;
+import com.symphony.bdk.core.config.model.BdkConfig;
+
 import com.fasterxml.jackson.databind.DeserializationFeature;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.json.JsonMapper;
-import com.symphony.bdk.core.config.legacy.LegacyConfigMapper;
-import com.symphony.bdk.core.config.legacy.model.LegacySymConfig;
-import com.symphony.bdk.core.config.model.BdkConfig;
-import com.symphony.bdk.core.config.exceptions.BdkConfigException;
 import lombok.extern.slf4j.Slf4j;
 
 import java.io.File;
