+++ symphony-bdk-core/src/main/java/com/symphony/bdk/core/config/BdkConfigParser.java
@@ -1,11 +1,12 @@
 package com.symphony.bdk.core.config;
 
+import com.symphony.bdk.core.config.exception.BdkConfigException;
+
 import com.fasterxml.jackson.databind.DeserializationFeature;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.json.JsonMapper;
 import com.fasterxml.jackson.dataformat.yaml.YAMLMapper;
-import com.symphony.bdk.core.config.exceptions.BdkConfigException;
 import lombok.extern.slf4j.Slf4j;
 import org.apiguardian.api.API;
