+++ symphony-bdk-examples/bdk-core-examples/src/main/java/com/symphony/bdk/examples/AuthMain.java
@@ -4,7 +4,7 @@ import com.symphony.bdk.core.SymphonyBdk;
 import com.symphony.bdk.core.auth.exception.AuthInitializationException;
 import com.symphony.bdk.core.auth.obo.Obo;
 import com.symphony.bdk.core.config.BdkConfigLoader;
-import com.symphony.bdk.core.exceptions.BdkConfigException;
+import com.symphony.bdk.core.config.exception.BdkConfigException;
 import com.symphony.bdk.gen.api.model.V4Message;
 
 import lombok.extern.slf4j.Slf4j;
